API/requirements.txt
anyio==3.6.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
h11==0.14.0
idna==3.4
imageio==2.22.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pycparser==2.21
pydantic==1.10.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.9.3
sniffio==1.3.0
starlette==0.20.4
tifffile==2022.10.10
typing_extensions==4.4.0
Werkzeug==2.2.2
